// schemas/model/event.ts
/**
 * Generated by orval v6.23.0 🍺
 * Do not edit manually.
 * My API
 * OpenAPI spec version: 1.0.0
 */
import type { EventExtraFields } from './eventExtraFields';

export interface Event {
  createdAt: string;
  displayName: string;
  extraFields?: EventExtraFields;
  icon: string;
  id: number;
  name: string;
  updatedAt: string;
}