// schemas/model/postApiUpload200.ts
/**
 * Generated by orval v6.23.0 🍺
 * Do not edit manually.
 * My API
 * OpenAPI spec version: 1.0.0
 */
import type { PostApiUpload200Bucket } from './postApiUpload200Bucket';

export type PostApiUpload200 = {
  bucket: PostApiUpload200Bucket;
  bucketName: string;
  createdAt: string;
  hash: string;
  id: number;
  mediaType: string;
  thumbnailHash: string;
  updatedAt: string;
};